+++ src/avsConfig.h
@@ -12,6 +12,12 @@ typedef uint16_t valueid_t; //type for Value ID
 typedef uint8_t vatid_t;    //type for Value Allocation Table ID
 typedef uint8_t vtail_t;    //type for tail length of Value
 
+/* define AVS_USE_ARDUINO_STRING_METHODS if you want use methods with Arduino String https://www.arduino.cc/reference/en/language/variables/data-types/stringobject/:
+  - cavsStorage::[insert, select, update] 
+  - cavsValue::[read, write]
+Please note that this will increase firmware size to 2 kb minimum */
+//#define AVS_USE_ARDUINO_STRING_METHODS
+
 //VAT = Value Allocation Table, contains singly linked lists of sector addresses(byte)
 //DS  = Data Storage, divided by sectors with AVS_DS_SECTOR_SIZE bytes, contains value data