+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This file contains changelog of [Arduino Value Storage Library](https://github.c
 
 
 ## [Unreleased]
+### Changed
+- Add AVS_USE_ARDUINO_STRING_METHODS in [avsConfig.h]. If defined: will declared methods, which are using [Arduino String](https://www.arduino.cc/reference/en/language/variables/data-types/stringobject/):
+  - [cavsStorage]::[insert, select, update] 
+  - [cavsValue]::[read, write]
 
 ## [0.9.1-alpha] - 2018-06-15
 ### Added
@@ -20,7 +24,7 @@ This file contains changelog of [Arduino Value Storage Library](https://github.c
   - [cavsValue]::[read, write]
 - [main.cpp] - add avs_test_5_String_methods() 
 ### Changed
-- Remake storage-config-selector section in avsConfig.h
+- Remake storage-config-selector section in [avsConfig.h]
 - [cavsStorage]::free(), can catch value_id = AVS_NOTSET_VALUE_ID
 ### Fixed
 - Fix bug in [cavsStorage]::free(), which was produce randomized effects.
@@ -37,3 +41,4 @@ This file contains changelog of [Arduino Value Storage Library](https://github.c
 [cavsStorage]: ./src/cavsStorage.h
 [cavsValue]: ./src/cavsValue.h
 [main.cpp]: ./src/main.cpp
+[avsConfig.h]: ./src/avsConfig.h